corrected mistake in utils.cpp

+++ Project3/src/utils.cpp
@@ -2,7 +2,7 @@
 #include<utils.hpp>
 
 // Solve the problem for one particle..
-arma::mat solve_analytical_1p(double v0, double x0, double z0, double tmax, double steps, double q, 
+arma::mat solve_analytical_1p(double v0, double x0, double z0, double tmax, int steps, int q, 
                               double B0, double V0, double m, double d) {
     arma::mat motion_r = arma::mat(steps, 3, arma::fill::zeros);
